+++ decidim-core/app/models/decidim/action_log.rb
@@ -19,7 +19,8 @@ module Decidim
                class_name: "Decidim::Component"
 
     belongs_to :resource,
-               polymorphic: true
+               polymorphic: true,
+               optional: true
 
     belongs_to :participatory_space,
                optional: true,
@@ -29,7 +30,8 @@ module Decidim
                optional: true,
                class_name: "PaperTrail::Version"
 
-    validates :organization, :user, :action, :resource, presence: true
+    validates :organization, :user, :action, presence: true
+    validates :resource, presence: true, if: ->(log) { log.action != "delete" }
 
     # To ensure records can't be deleted
     before_destroy { |_record| raise ActiveRecord::ReadOnlyRecord }