+++ decidim-sortitions/db/migrate/20180102100101_add_author_to_sortitions.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddAuthorToSortitions < ActiveRecord::Migration[5.1]
+  def change
+    add_reference :decidim_module_sortitions_sortitions, :decidim_author, index: true
+  end
+end