+++ decidim-surveys/app/forms/decidim/surveys/admin/survey_question_form.rb
@@ -22,7 +22,7 @@ module Decidim
 
         def map_model(model)
           self.answer_options = model.answer_options.each_with_index.map do |option, id|
-            SurveyQuestionAnswerOptionForm.new(option.merge(id: id + 1))
+            SurveyQuestionAnswerOptionForm.new(option.merge(id: id + 1, deleted: false))
           end
         end